Rename styles/config to styles/theme, update usage references, clean that file up
