+++ src/content-single/EventContentItem/components/EventPanel.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import styled, { withTheme } from 'styled-components/macro';
 
-import { baseUnit } from 'styles/config';
+import { baseUnit } from 'styles/theme';
 
 import { Icon } from 'ui/Icons';
